requirements.txt
flask==2.3.0
mistralai==0.4.2
requests==2.31.0
markdown==3.4.3
python-docx==0.8.11
pypdf==3.11.0   